app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
injector
aiohttp
ddgs
playwright
langchain
langchain-community
langchain_anthropic

# Testing dependencies
pytest
pytest-cov
pytest-asyncio
Faker
langdetect
pandas